+++ src/gallery-sm-treeview/js/treeview.js
@@ -14,13 +14,16 @@ TreeView widget.
 @constructor
 @extends View
 @uses Tree
+@uses Tree.Labelable
 @uses Tree.Openable
 @uses Tree.Selectable
 **/
 
 var getClassName = Y.ClassNameManager.getClassName,
 
-TreeView = Y.Base.create('treeView', Y.View, [Y.Tree, Y.Tree.Openable, Y.Tree.Selectable], {
+TreeView = Y.Base.create('treeView', Y.View, [
+    Y.Tree, Y.Tree.Labelable, Y.Tree.Openable, Y.Tree.Selectable
+], {
     // -- Public Properties ----------------------------------------------------
 
     /**